+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Features
 
 -   `[melody-parser]`: Add parsing of declarations like `<!DOCTYPE html>`, along with new Parser option `ignoreDeclarations`.
+-   `[melody-parser]`: Fix bug where `1..5` results in a `MemberExpression` instead of a `BinaryRangeExpression`
 
 ### Bug fixes